refactor: update imports to use secrets_sanitizer and clean up unused serialization code in Log class
